refactor(connect): move cardano method definitions to corresponding folder
