+++ packages/connect-common/src/types/api/cardano/cardanoGetAddress.ts
@@ -1,5 +1,5 @@
-import type { BundledParams, Params, Response } from '../params';
-import type { CardanoAddress, CardanoGetAddress } from './cardano';
+import type { CardanoAddress, CardanoGetAddress } from './common';
+import type { BundledParams, Params, Response } from '../../params';
 
 export declare function cardanoGetAddress(
     params: Params<CardanoGetAddress>,
