+++ packages/connect-common/src/types/api/cardano/cardanoSignMessage.ts
@@ -1,0 +1,6 @@
+import type { CardanoSignMessage, CardanoSignedMessage } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function cardanoSignMessage(
+    params: Params<CardanoSignMessage>,
+): Response<CardanoSignedMessage>;
