--- packages/connect-common/src/types/api/cardanoSignMessage.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { CardanoSignMessage, CardanoSignedMessage } from './cardano';
-
-export declare function cardanoSignMessage(
-    params: Params<CardanoSignMessage>,
-): Response<CardanoSignedMessage>;
